requirements.txt
yfinance>=0.2.18
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
requests>=2.31.0
hyperliquid-python-sdk>=0.0.15
eth-account>=0.9.0
python-dotenv>=1.0.0
schedule>=1.2.0 